Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/xhr/send-entity-body-get-head-async-expected.txt


PASS XMLHttpRequest: send() - non-empty data argument and GET/HEAD - async, no upload events should fire (GET) 
FAIL XMLHttpRequest: send() - non-empty data argument and GET/HEAD - async, no upload events should fire (HEAD) assert_equals: expected "NO" but got "0"
